Move util files to new file util.py
