+++ indra_depmap_service/api.py
@@ -1,3 +1,4 @@
+"""INDRA Causal Network Search API"""
 import os
 import json
 import pickle
@@ -5,43 +6,26 @@ import logging
 import argparse
 from sys import argv
 from os import path, makedirs
-from datetime import datetime
 from time import time, gmtime, strftime
 
-import boto3
-import networkx as nx
-from botocore import UNSIGNED
-from botocore.client import Config
 from jinja2 import Template
 from flask import Flask, request, abort, Response, render_template
-from indra_db.util.dump_sif import load_db_content, make_dataframe, NS_LIST
+
 from indra.config import CONFIG_DICT
-from indra.statements import get_all_descendants, Activation, Inhibition, \
-    IncreaseAmount, DecreaseAmount, AddModification, RemoveModification
-from depmap_analysis.network_functions import net_functions as nf
 from depmap_analysis.network_functions.indra_network import IndraNetwork
 from depmap_analysis.util.io_functions import pickle_open, dump_it_to_pickle
+
+from .util import load_indra_graph, _get_s3_client,\
+    get_queryable_stmt_types, API_PATH as HERE, CACHE, INDRA_DG, \
+    INDRA_DG_CACHE, INDRA_SEG, INDRA_SEG_CACHE, INDRA_SNG_CACHE, TEST_DG_CACHE
 
 app = Flask(__name__)
 app.config['DEBUG'] = True
 
 logger = logging.getLogger('INDRA Network Search API')
 
-HERE = path.dirname(path.abspath(__file__))
-CACHE = path.join(HERE, '_cache')
-STATIC = path.join(HERE, 'static')
 S3_BUCKET = 'depmap-analysis'
-INDRA_MDG = 'indranet_multi_digraph_latest.pkl'
-INDRA_DG = 'indranet_dir_graph_latest.pkl'
-INDRA_SNG = 'indranet_sign_node_graph_latest.pkl'
-INDRA_SEG = 'indranet_sign_edge_graph_latest.pkl'
 
-TEST_MDG_CACHE = path.join(CACHE, 'test_mdg_network.pkl')
-INDRA_MDG_CACHE = path.join(CACHE, INDRA_MDG)
-TEST_DG_CACHE = path.join(CACHE, 'test_dir_network.pkl')
-INDRA_DG_CACHE = path.join(CACHE, INDRA_DG)
-INDRA_SNG_CACHE = path.join(CACHE, INDRA_SNG)
-INDRA_SEG_CACHE = path.join(CACHE, INDRA_SEG)
 
 FILES = {
     'dir_graph_path': INDRA_DG_CACHE if path.isfile(INDRA_DG_CACHE)
@@ -71,18 +55,6 @@ MAX_PATHS = 50
 TIMEOUT = 30  # Timeout in seconds
 
 
-def _todays_date():
-    return datetime.now().strftime('%Y%m%d')
-
-
-def _get_s3_client(unsigned=True):
-    if unsigned:
-        return boto3.client('s3',
-                            config=Config(signature_version=UNSIGNED))
-    else:
-        return boto3.client('s3')
-
-
 # Create a template object from the template file, load once
 def _load_template(fname):
     template_path = path.join(HERE, fname)
@@ -90,69 +62,6 @@ def _load_template(fname):
         template_str = f.read()
         template = Template(template_str)
     return template
-
-
-# Copied from emmaa_service/api.py
-def get_queryable_stmt_types():
-    """Return Statement class names that can be used for querying."""
-    def get_sorted_descendants(cls):
-        return sorted(get_names(get_all_descendants(cls)))
-
-    def get_names(classes):
-        return [s.__name__ for s in classes]
-
-    stmt_types = \
-        get_names([Activation, Inhibition, IncreaseAmount, DecreaseAmount]) + \
-        get_sorted_descendants(AddModification) + \
-        get_sorted_descendants(RemoveModification)
-    return stmt_types
-
-
-def load_indra_graph(dir_graph_path, multi_digraph_path=None,
-                     sign_node_graph_path=None, sign_edge_graph_path=None,
-                     update=False, belief_dict=None, strat_ev_dict=None,
-                     include_entity_hierarchies=True, verbosity=0):
-    """Return a nx.DiGraph and nx.MultiDiGraph representation an INDRA DB dump
-
-    If update is True, make a fresh snapshot from the INDRA DB.
-    WARNING: this typically requires a lot of RAM and might slow down your
-    system significantly.
-    """
-    global INDRA_DG_CACHE, INDRA_MDG_CACHE, INDRA_SNG_CACHE, INDRA_SEG_CACHE
-    indra_multi_digraph = nx.MultiDiGraph()
-    indra_signed_edge_graph = nx.MultiDiGraph()
-    indra_signed_node_graph = nx.DiGraph()
-
-    if update:
-        df = make_dataframe(True, load_db_content(True, NS_LIST))
-        options = {'df': df,
-                   'belief_dict': belief_dict,
-                   'strat_ev_dict': strat_ev_dict,
-                   'include_entity_hierarchies': include_entity_hierarchies,
-                   'verbosity': verbosity}
-        indra_dir_graph = nf.sif_dump_df_to_nx_digraph(**options)
-        dump_it_to_pickle(dir_graph_path, indra_dir_graph)
-        INDRA_DG_CACHE = path.join(CACHE, dir_graph_path)
-        if multi_digraph_path:
-            indra_multi_digraph = nf.sif_dump_df_to_nx_digraph(
-                graph_type='multidigraph', **options)
-            dump_it_to_pickle(multi_digraph_path, indra_multi_digraph)
-            INDRA_MDG_CACHE = path.join(CACHE, multi_digraph_path)
-        if sign_node_graph_path or sign_edge_graph_path:
-            indra_signed_edge_graph, indra_signed_node_graph = \
-                nf.sif_dump_df_to_nx_digraph(graph_type='signed', **options)
-    else:
-        logger.info('Loading indra network representations from pickles')
-        indra_dir_graph = pickle_open(dir_graph_path)
-        if multi_digraph_path:
-            indra_multi_digraph = pickle_open(multi_digraph_path)
-        if sign_edge_graph_path:
-            indra_signed_edge_graph = pickle_open(sign_edge_graph_path)
-        if sign_node_graph_path:
-            indra_signed_node_graph = pickle_open(sign_node_graph_path)
-        logger.info('Finished loading indra networks.')
-    return indra_dir_graph, indra_multi_digraph, indra_signed_edge_graph,\
-        indra_signed_node_graph
 
 
 # Load network
