+++ indra_depmap_service/util.py
@@ -1,0 +1,111 @@
+"""Utility functions for the INDRA Causal Network Search API in api.py"""
+import boto3
+import logging
+import networkx as nx
+from os import path
+from botocore import UNSIGNED
+from datetime import datetime
+from botocore.client import Config
+
+from indra_db.util.dump_sif import load_db_content, make_dataframe, NS_LIST
+from indra.statements import get_all_descendants, Activation, Inhibition, \
+    IncreaseAmount, DecreaseAmount, AddModification, RemoveModification
+
+from depmap_analysis.network_functions import net_functions as nf
+from depmap_analysis.util.io_functions import pickle_open, dump_it_to_pickle
+
+
+logger = logging.getLogger('INDRA Network Search util')
+
+API_PATH = path.dirname(path.abspath('./api.py'))
+CACHE = path.join(API_PATH, '_cache')
+STATIC = path.join(API_PATH, 'static')
+
+INDRA_MDG = 'indranet_multi_digraph_latest.pkl'
+INDRA_DG = 'indranet_dir_graph_latest.pkl'
+INDRA_SNG = 'indranet_sign_node_graph_latest.pkl'
+INDRA_SEG = 'indranet_sign_edge_graph_latest.pkl'
+
+TEST_MDG_CACHE = path.join(CACHE, 'test_mdg_network.pkl')
+INDRA_MDG_CACHE = path.join(CACHE, INDRA_MDG)
+TEST_DG_CACHE = path.join(CACHE, 'test_dir_network.pkl')
+INDRA_DG_CACHE = path.join(CACHE, INDRA_DG)
+INDRA_SNG_CACHE = path.join(CACHE, INDRA_SNG)
+INDRA_SEG_CACHE = path.join(CACHE, INDRA_SEG)
+
+
+def _todays_date():
+    return datetime.now().strftime('%Y%m%d')
+
+
+def _get_s3_client(unsigned=True):
+    if unsigned:
+        return boto3.client('s3',
+                            config=Config(signature_version=UNSIGNED))
+    else:
+        return boto3.client('s3')
+
+
+# Copied from emmaa_service/api.py
+def get_queryable_stmt_types():
+    """Return Statement class names that can be used for querying."""
+    def _get_sorted_descendants(cls):
+        return sorted(_get_names(get_all_descendants(cls)))
+
+    def _get_names(classes):
+        return [s.__name__ for s in classes]
+
+    stmt_types = \
+        _get_names([
+            Activation, Inhibition, IncreaseAmount, DecreaseAmount
+        ]) + \
+        _get_sorted_descendants(AddModification) + \
+        _get_sorted_descendants(RemoveModification)
+    return stmt_types
+
+
+def load_indra_graph(dir_graph_path, multi_digraph_path=None,
+                     sign_node_graph_path=None, sign_edge_graph_path=None,
+                     update=False, belief_dict=None, strat_ev_dict=None,
+                     include_entity_hierarchies=True, verbosity=0):
+    """Return a nx.DiGraph and nx.MultiDiGraph representation an INDRA DB dump
+
+    If update is True, make a fresh snapshot from the INDRA DB.
+    WARNING: this typically requires a lot of RAM and might slow down your
+    system significantly.
+    """
+    global INDRA_DG_CACHE, INDRA_MDG_CACHE, INDRA_SNG_CACHE, INDRA_SEG_CACHE
+    indra_multi_digraph = nx.MultiDiGraph()
+    indra_signed_edge_graph = nx.MultiDiGraph()
+    indra_signed_node_graph = nx.DiGraph()
+
+    if update:
+        df = make_dataframe(True, load_db_content(True, NS_LIST))
+        options = {'df': df,
+                   'belief_dict': belief_dict,
+                   'strat_ev_dict': strat_ev_dict,
+                   'include_entity_hierarchies': include_entity_hierarchies,
+                   'verbosity': verbosity}
+        indra_dir_graph = nf.sif_dump_df_to_nx_digraph(**options)
+        dump_it_to_pickle(dir_graph_path, indra_dir_graph)
+        INDRA_DG_CACHE = path.join(CACHE, dir_graph_path)
+        if multi_digraph_path:
+            indra_multi_digraph = nf.sif_dump_df_to_nx_digraph(
+                graph_type='multidigraph', **options)
+            dump_it_to_pickle(multi_digraph_path, indra_multi_digraph)
+            INDRA_MDG_CACHE = path.join(CACHE, multi_digraph_path)
+        if sign_node_graph_path or sign_edge_graph_path:
+            indra_signed_edge_graph, indra_signed_node_graph = \
+                nf.sif_dump_df_to_nx_digraph(graph_type='signed', **options)
+    else:
+        logger.info('Loading indra network representations from pickles')
+        indra_dir_graph = pickle_open(dir_graph_path)
+        if multi_digraph_path:
+            indra_multi_digraph = pickle_open(multi_digraph_path)
+        if sign_edge_graph_path:
+            indra_signed_edge_graph = pickle_open(sign_edge_graph_path)
+        if sign_node_graph_path:
+            indra_signed_node_graph = pickle_open(sign_node_graph_path)
+        logger.info('Finished loading indra networks.')
+    return indra_dir_graph, indra_multi_digraph, indra_signed_edge_graph,\
+        indra_signed_node_graph
